requirements.txt
# Core Dependencies
python-telegram-bot==20.7
SQLAlchemy==2.0.25
APScheduler==3.10.4
scikit-learn==1.3.2
redis==5.0.1
python-dotenv==1.0.0

# Development Dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.2
pre-commit==3.5.0

# Database
alembic==1.12.1
psycopg2-binary==2.9.9  # For future PostgreSQL migration

# Monitoring
sentry-sdk==1.34.0
prometheus-client==0.19.0

# Testing
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0 